Update based on model changes
