+++ app/viewmodels/checkAnswers/summary/ApplicantSummary.scala
@@ -22,7 +22,7 @@ import play.api.i18n.Messages
 import uk.gov.hmrc.govukfrontend.views.viewmodels.summarylist._
 
 import models.UserAnswers
-import models.requests.{ApplicationRequest, IndividualApplicant, OrganisationApplicant}
+import models.requests.ApplicationRequest
 import viewmodels.checkAnswers._
 import viewmodels.govuk.summarylist._
 
@@ -33,13 +33,7 @@ sealed trait ApplicantSummary {
 
 object ApplicantSummary {
   def apply(applicationRequest: ApplicationRequest)(implicit messages: Messages) =
-    applicationRequest.applicant match {
-      case IndividualApplicant(_)      =>
-        CheckRegisteredDetailsSummary.rows(applicationRequest).map(_.copy(actions = None))
-      case OrganisationApplicant(_, _) =>
-        CheckRegisteredDetailsForAgentsSummary.rows(applicationRequest).map(_.copy(actions = None))
-    }
-
+    CheckRegisteredDetailsSummary.rows(applicationRequest).map(_.copy(actions = None))
 }
 
 case class IndividualApplicantSummary(rows: SummaryList) extends ApplicantSummary {
@@ -61,13 +55,11 @@ case class AgentSummary(rows: SummaryList) extends ApplicantSummary {
 }
 
 object AgentSummary {
-  def apply(applicationRequest: ApplicationRequest)(implicit messages: Messages) =
-    CheckRegisteredDetailsForAgentsSummary.rows(applicationRequest).map(_.copy(actions = None))
-
   def apply(userAnswers: UserAnswers)(implicit messages: Messages): AgentSummary = {
 
     val contactDetailsRows = BusinessContactDetailsSummary.rows(userAnswers).orEmpty
     val roleRow            = AgentRoleSummary.row(userAnswers).orEmpty
+
     AgentSummary(SummaryListViewModel(contactDetailsRows ++ roleRow))
   }
 }
